feat: move command package under dto file
