+++ src/main/kotlin/org/appjam/smashing/domain/auth/dto/request/SignInRequest.kt
@@ -1,7 +1,7 @@
 package org.appjam.smashing.domain.auth.dto.request
 
 import jakarta.validation.constraints.NotBlank
-import org.appjam.smashing.domain.auth.command.SignInRequestCommand
+import org.appjam.smashing.domain.auth.dto.command.SignInRequestCommand
 
 data class SignInRequest(
     @field:NotBlank(message = "엑세스 토큰을 입력해주세요.")
