+++ src/main/kotlin/org/appjam/smashing/domain/auth/dto/request/SignUpRequest.kt
@@ -1,7 +1,7 @@
 package org.appjam.smashing.domain.auth.dto.request
 
 import jakarta.validation.constraints.NotBlank
-import org.appjam.smashing.domain.auth.command.SignUpRequestCommand
+import org.appjam.smashing.domain.auth.dto.command.SignUpRequestCommand
 import org.appjam.smashing.domain.user.enums.Gender
 import org.appjam.smashing.global.common.validator.annotation.ValidEnum
 import org.appjam.smashing.global.extensions.ofIgnoreCase
